requirements.txt
matplotlib>=3.7
numpy>=1.24
Pillow>=10.0
tqdm>=4.66

